% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larinf.R
\name{larinf}
\alias{larinf}
\title{Least angle regression inference}
\usage{
larinf(X, y, rescale_y = T, B = 500, alpha = 0.05, m_bar = NULL)
}
\arguments{
\item{X}{a design matrix.}

\item{y}{a response vector.}

\item{rescale_y}{a logical. If \code{TRUE} then \code{y} is scaled by one over root n.}

\item{B}{the number of Monte Carlo draws for approximating the bootstrap distribution.}

\item{alpha}{the significance level.}

\item{m_bar}{optional argument fixing the number of nonzero entrance correlations at which to threshold. By default the function estimates this with \code{mest()}.}
}
\value{
The \code{larinf()} function returns an object of class \code{larinf}. An object of class \code{larinf} is a list containing:

\describe{
\item{\code{lob}}{a matrix with each column giving the lower bound of the confidence interval for the step coefficients at the corresponding step.}
\item{\code{upb}}{a matrix with each column giving the upper bound of the confidence interval for the step coefficients at the corresponding step.}
\item{\code{b_bar}}{a matrix with each column giving the estimated step coefficients at a step. The number of columns is equal to the estimated number of steps in the population path, and the last column contains the coefficients corresponding to the orthogonal projection of the response on the span of the columns in the active set.}
\item{\code{loC}}{a vector containing the lower limits of the confidence intervals for the step correlations.}
\item{\code{upC}}{a vector containing the upper limits of the confidence intervals for the step correlations.}
\item{\code{sigma_hat}}{the estimate of the error standard deviation.}
\item{\code{lar_out}}{an object of class \code{lar} returned by a call to the \code{lar()} function.}
\item{\code{mest_out}}{an object of class \code{mest} returned by a call to the \code{mest()} function.}
\item{\code{actprob}}{a matrix giving the proportion of bootstrap samples on which each variable was active at each step.}
\item{\code{B}}{the number of bootstrap samples drawn.}
\item{\code{alpha}}{the significance level.}
\item{\code{bm_lo}}{lower limits of the classical t-distribution-based confidence intervals for the regression coefficients in the model at the estimated number of steps. }
\item{\code{bm_up}}{upper limits of the classical t-distribution-based confidence intervals for the regression coefficients in the model at the estimated number of steps. }
}
}
\description{
Computes bootstrap confidence intervals for least angle regression entrance correlations.
}
\details{
The \code{larinf()} function centers the response vector \code{y} and centers and normalizes the columns of \code{X}.
}
\examples{
# set parameters for generating data
n <- 100
p <- 10
sigma <- 1
b <- c(1,-1,2,-1.5,rep(0,p-4))
Sigma <- (1/2)^abs(outer(1:p,1:p,FUN = "-"))
chol_Sigma <- chol(Sigma)

# generate X and y
X <- matrix(rnorm(n*p),n,p) \%*\% chol_Sigma
mu <- drop(X \%*\% b)
e <- rnorm(n,0,sigma)
y <- mu + e

# perform least angle regression inference
larinf_out<- larinf(X,y)
plot(larinf_out)
larinf_out
}
\references{
Gregory, K. and Nordman, D. (2025+) Least angle regression inference.
}
\seealso{
\code{\link[=plot.larinf]{plot.larinf()}} for plotting and \code{\link[=print.larinf]{print.larinf()}} for printing the results of least angle regression inference. See \code{\link[=mest]{mest()}} for selecting the number of nonzero entrance correlations at which to threshold for the bootstrap.
}
\author{
Karl Gregory
}
