% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larinf.R
\name{plot.lar}
\alias{plot.lar}
\title{Plot method for class \code{lar}}
\usage{
\method{plot}{lar}(x, omaadd = c(0, 0, 0, 0), madd = c(0, 0, 0, 0), m = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{lar}, usually the result of a call to \code{lar}. The LAR path may not have tied entrances.}

\item{omaadd}{a vector of length 4 giving amounts to add to each of the four outer margins (which may be necessary to include longer variable names).}

\item{madd}{a vector of length 4 giving amounts to add to each of the four inner margins (which may be necessary to include longer variable names).}

\item{m}{the number of steps out to which to plot the LAR path}

\item{...}{additional arguments.}
}
\description{
Plot the steps of the LAR path based on the output of the \code{lar} function.
}
\examples{
# set parameters for generating data
n <- 100
p <- 10
sigma <- 1
b <- c(1,-1,2,-1.5,rep(0,p-4))
Sigma <- (1/2)^abs(outer(1:p,1:p,FUN = "-"))
chol_Sigma <- chol(Sigma)

# generate X and y
X <- matrix(rnorm(n*p),n,p) \%*\% chol_Sigma
mu <- drop(X \%*\% b)
e <- rnorm(n,0,sigma)
y <- mu + e

# compute population path and sample path
lar_mu <- lar(X,mu)
lar_y <- lar(X,y)

# plot population path and sample path
plot(lar_mu)
plot(lar_y)

}
\author{
Karl Gregory
}
