% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larinf.R
\name{plot.larinf}
\alias{plot.larinf}
\title{Plot method for class \code{larinf}}
\usage{
\method{plot}{larinf}(x, omaadd = c(0, 0, 0, 0), ...)
}
\arguments{
\item{x}{an object of class \code{larinf}, usuall the result of a call to \code{larinf}.}

\item{omaadd}{a vector of length 4 giving amounts to add to each of the four outer margins (which may be necessary to include longer variable names).}

\item{...}{additional arguments.}
}
\description{
Make a cobweb plot of the least angle regression path and display confidence intervals for the entrance correlations as well as for the regression coefficients.
}
\examples{
# set parameters for generating data
n <- 100
p <- 10
sigma <- 1
b <- c(1,-1,2,-1.5,rep(0,p-4))
Sigma <- (1/2)^abs(outer(1:p,1:p,FUN = "-"))
chol_Sigma <- chol(Sigma)

# generate X and y
X <- matrix(rnorm(n*p),n,p) \%*\% chol_Sigma
mu <- drop(X \%*\% b)
e <- rnorm(n,0,sigma)
y <- mu + e

# perform least angle regression inference
larinf_out<- larinf(X,y)
plot(larinf_out)

}
\author{
Karl Gregory
}
