% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larinf.R
\name{plot.mest}
\alias{plot.mest}
\title{Plot method for class \code{mest}}
\usage{
\method{plot}{mest}(x, ...)
}
\arguments{
\item{x}{an object of class \code{mest}, usually the result of a call to \code{mest}.}

\item{...}{further arguments.}
}
\description{
Make plots related to the selection of the number of nonzero entrance correlations at which to threshold.
}
\examples{
# set parameters for generating data
n <- 100
p <- 10
sigma <- 1
beta <- c(1,-1,2,-1.5,rep(0,p-4))
Sigma <- (1/2)^abs(outer(1:p,1:p,FUN = "-"))
chol_Sigma <- chol(Sigma)

# generate data
X <- scale(matrix(rnorm(n*p),n,p) \%*\% chol_Sigma)*sqrt(n/(n-1))
mu <- drop(X \%*\% beta)
e <- rnorm(n,0,sigma)
y <- mu + e - mean(e)

# select the number of nonzero entrance correlations at which to threshold
mest_out <- mest(X,y)
plot(mest_out)

}
\author{
Karl Gregory
}
