% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larinf.R
\name{print.larinf}
\alias{print.larinf}
\title{Print method for class \code{larinf}}
\usage{
\method{print}{larinf}(x, ...)
}
\arguments{
\item{x}{an object of class \code{larinf}, usuall the result of a call to \code{larinf}.}

\item{...}{further arguments.}
}
\description{
Print a table of estimated entrance correlations with bootstrap confidence intervals.
}
\examples{
# set parameters for generating data
n <- 100
p <- 10
sigma <- 1
b <- c(1,-1,2,-1.5,rep(0,p-4))
Sigma <- (1/2)^abs(outer(1:p,1:p,FUN = "-"))
chol_Sigma <- chol(Sigma)

# generate X and y
X <- matrix(rnorm(n*p),n,p) \%*\% chol_Sigma
mu <- drop(X \%*\% b)
e <- rnorm(n,0,sigma)
y <- mu + e

# perform least angle regression inference
larinf_out<- larinf(X,y)
print(larinf_out)

}
\author{
Karl Gregory
}
