% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larinf.R
\name{lar}
\alias{lar}
\title{Least angle regression}
\usage{
lar(X, y, rescale_y = T)
}
\arguments{
\item{X}{a design matrix.}

\item{y}{a response vector.}

\item{rescale_y}{a logical. If \code{TRUE} then \code{y} is scaled by one over root n.}
}
\value{
The \code{lar()} function returns an object of class \code{lar}. An object of class \code{lar} is a list containing:

\describe{
\item{\code{a}}{A matrix containing the sequence of equiangular vectors.}
\item{\code{mu}}{The value of the terminal predictor.}
\item{\code{A}}{A vector containing the sequence of angles.}
\item{\code{C}}{A vector containing the absolute entrance correlations.}
\item{\code{g}}{A vector containing the sequence of gamma values.}
\item{\code{b}}{A matrix containing the sequence of regression coefficient vectors.}
\item{\code{th}}{A vector containing the entrance correlations for the columns of \code{X}.}
\item{\code{ord}}{A vector giving the order in which the columns of \code{X} entered the active set.}
\item{\code{nnew}}{A vector giving the number of the columns of \code{X} entering on each step.}
\item{\code{varnames}}{The names of the columns of \code{X} if non-null.}
\item{\code{ehat}}{The residuals.}
\item{\code{rescale_y}}{The argument given for \code{rescale_y}.}
}
}
\description{
Runs the least angle regression algorithm.
}
\details{
The \code{lar()} function centers the response vector \code{y} and centers the columns of \code{X} and scales them to have
unit norm before executing the least angle regression algorithm.  The function can accommodate tied entrances.
}
\examples{
# set parameters for generating data
n <- 100
p <- 10
sigma <- 1
b <- c(1,-1,2,-1.5,rep(0,p-4))
Sigma <- (1/2)^abs(outer(1:p,1:p,FUN = "-"))
chol_Sigma <- chol(Sigma)

# generate X and y
X <- matrix(rnorm(n*p),n,p) \%*\% chol_Sigma
mu <- drop(X \%*\% b)
e <- rnorm(n,0,sigma)
y <- mu + e

# compute noiseless path and sample path
lar_mu <- lar(X,mu)
lar_y <- lar(X,y)

# plot noiseless path and overlay sample path
plot(lar_mu)
plot(lar_y,add = TRUE, lty = 3)
}
\references{
Efron B., Hastie, T., Johnstone, I., and Tibshirani, R. (2004) Least angle regression.
\emph{Annals of Statistics}, \strong{32(2)}: 407-499. \href{https://doi.org/10.1214/009053604000000067}{doi:10.1214/009053604000000067}

Gregory, K. and Nordman, D. (2025+) Least angle regression inference.
}
\seealso{
\code{\link[=plot.lar]{plot.lar()}} for plotting the least angle regression path.
}
\author{
Karl Gregory
}
