% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larinf.R
\name{larinf}
\alias{larinf}
\title{Least angle regression inference}
\usage{
larinf(X, y, B = 500, alpha = 0.05, thresh = 0.8, ncp = 0.1, m_bar = NULL)
}
\arguments{
\item{X}{a design matrix with columns centered to have mean zero.}

\item{y}{a response vector centered to have mean zero.}

\item{B}{the number of Monte Carlo draws for approximating the bootstrap distribution.}

\item{alpha}{the significance level.}

\item{thresh}{an argument passed to the \code{mest} function.}

\item{ncp}{an argument passed to the \code{mest} function.}

\item{m_bar}{optional argument fixing the number of nonzero entrance correlations at which to threshold.}
}
\value{
\code{larinf} returns an object of class \code{larinf}. An object of class \code{larinf} is a list containing:

\describe{
\item{\code{lo}}{The lower bounds of the entrance correlation confidence intervals.}
\item{\code{up}}{The upper bounds of the entrance correlation confidence intervals.}
\item{\code{rej}}{An indicator of whether the entrance correlation confidence intervals exclude zero.}
\item{\code{b_lo}}{The lower bounds of the regression coefficient confidence intervals.}
\item{\code{b_up}}{The upper bounds of the regression coefficient confidence intervals.}
\item{\code{b_rej}}{An indicator of whether the regression coefficient confidence intervals exclude zero.}
\item{\code{sigma_hat}}{The estimator of the error term standard deviation.}
\item{\code{lar_out}}{The object returned by \code{lar(X,y)}.}
\item{\code{mest_out}}{The object returned by \code{mest(X,y,thresh,ncp,m_bar)}.}
\item{\code{B}}{The number of Monte Carlo draws used to approximate the bootstrap distribution.}
\item{\code{alpha}}{The significance level.}
}
}
\description{
Least angle regression inference
}
\examples{
# set parameters for generating data
n <- 100
p <- 10
sigma <- 1
beta <- c(1,-1,2,-1.5,rep(0,p-4))
Sigma <- (1/2)^abs(outer(1:p,1:p,FUN = "-"))
chol_Sigma <- chol(Sigma)

# generate some data
X <- scale(matrix(rnorm(n*p),n,p) \%*\% chol_Sigma)*sqrt(n/(n-1))
mu <- drop(X \%*\% beta)
e <- rnorm(n,0,sigma)
y <- mu + e - mean(e)

# perform least angle regression inference
larinf_out <- larinf(X,y)
plot(larinf_out)
print(larinf_out)
}
\references{
Gregory, K. and Nordman, D. (2025+) Least angle regression inference. \emph{In progress}
}
\seealso{
\code{\link[=plot.larinf]{plot.larinf()}} for plotting and \code{\link[=print.larinf]{print.larinf()}} for printing the results of least angle regression inference and \code{\link[=mest]{mest()}} for selecting the number of nonzero entrance correlations at which to threshold for the bootstrap.
}
\author{
Karl Gregory
}
