% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larinf.R
\name{mest}
\alias{mest}
\title{Estimate the number of nonzero entrance correlations}
\usage{
mest(lar_out, thresh = 0.8, ncp = 0.1, m_bar = NULL)
}
\arguments{
\item{lar_out}{a \code{lar} object, such as returned by the \code{lar()} function.}

\item{thresh}{the value xi as described in Gregory and Nordman (2025+).}

\item{ncp}{the value c as described in Gregory and Nordman (2025+).}

\item{m_bar}{a user-specified number of nonzero entrance correlations.}
}
\value{
\code{mest} returns an object of class \code{mest}. An object of class \code{mest} is a list containing:

\describe{
\item{\code{m_bar}}{The selected number of nonzero entrance correlations.}
\item{\code{th_bar}}{A vector containing the thresholded entrance correlations.}
\item{\code{W}}{A vector containing the Wk values.}
\item{\code{thresh}}{The value of the \code{thresh} argument.}
\item{\code{ncp}}{The value of the \code{ncp} argument.}
\item{\code{Fval}}{The quantile of the F distribution to which the Wk are compared.}
}
}
\description{
Estimate the number of nonzero entrance correlations in the least angle regression path.
}
\examples{
# set parameters for generating data
n <- 100
p <- 10
sigma <- 1
b <- c(1,-1,2,-1.5,rep(0,p-4))
Sigma <- (1/2)^abs(outer(1:p,1:p,FUN = "-"))
chol_Sigma <- chol(Sigma)

# generate X and y
X <- matrix(rnorm(n*p),n,p) \%*\% chol_Sigma
mu <- drop(X \%*\% b)
e <- rnorm(n,0,sigma)
y <- mu + e

# estimate the number of nonzero entrance correlations
lar_out <- lar(X,y)
mest_out <- mest(lar_out)
plot(mest_out)
}
\references{
Gregory, K. and Nordman, D. (2025+) Least angle regression inference. \emph{In progress}
}
\seealso{
\code{\link[=plot.mest]{plot.mest()}} for producing plots related to the selection of the number of nonzero entrance correlations.
}
\author{
Karl Gregory
}
