% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larinf.R
\name{plot.lar}
\alias{plot.lar}
\title{Plot method for class \code{lar}}
\usage{
\method{plot}{lar}(
  x,
  main = "",
  add = FALSE,
  lty = 1,
  xlim = NULL,
  ylim = NULL,
  which_lab = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{lar}, usually the result of a call to \code{lar}.}

\item{main}{the main title of the plot.}

\item{add}{whether the cobweb plot should be added to the existing plot.}

\item{lty}{the line type.}

\item{xlim}{the limits of the horizontal axis if \code{add == FALSE}.}

\item{ylim}{the limits of the vertical axis if \code{add == FALSE}.}

\item{which_lab}{optional vector giving the indices of variables to be labeled.}

\item{...}{additional arguments.}
}
\description{
Make a cobweb plot of the least angle regression path based on the output of the \code{lar} function.
}
\examples{
# set parameters for generating data
n <- 100
p <- 10
sigma <- 1
b <- c(1,-1,2,-1.5,rep(0,p-4))
Sigma <- (1/2)^abs(outer(1:p,1:p,FUN = "-"))
chol_Sigma <- chol(Sigma)

# generate X and y
X <- matrix(rnorm(n*p),n,p) \%*\% chol_Sigma
mu <- drop(X \%*\% b)
e <- rnorm(n,0,sigma)
y <- mu + e

# compute noiseless path and sample path
lar_mu <- lar(X,mu)
lar_y <- lar(X,y)

# plot noiseless path and overlay sample path
plot(lar_mu)
plot(lar_y,add = TRUE, lty = 3)

}
\author{
Karl Gregory
}
