% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larinf.R
\name{plot.larinf}
\alias{plot.larinf}
\title{Plot method for class \code{larinf}}
\usage{
\method{plot}{larinf}(
  x,
  omaadd = c(0, 0, 0, 0),
  madd = c(0, 0, 0, 0),
  text = NULL,
  m = NULL,
  which = 1,
  ...
)
}
\arguments{
\item{x}{an object of class \code{larinf}, usually the result of a call to \code{larinf}.}

\item{omaadd}{a vector of length 4 giving amounts to add to each of the four outer margins (which may be necessary to include longer variable names).}

\item{madd}{a vector of length 4 giving amounts to add to each of the four inner margins (which may be necessary to include longer variable names).}

\item{text}{an optional argument giving text to be printed as a title on the plot}

\item{m}{an optional argument giving the number of steps out to which to plot the results of LAR inference. The default is the estimated number of steps from the \code{mest()} function.}

\item{which}{if equal to 1, the inferred LAR path is plotted; if equal to 2 diagnostic plots are plotted, i.e. the tail sum statistics and the bootstrap probability of active set membership at each step for each variable. Equal to 1 by default.}

\item{...}{additional arguments.}
}
\description{
Plot of the inferred LAR path out to the estimated number of steps.
}
\examples{
# set parameters for generating data
n <- 100
p <- 10
sigma <- 1
b <- c(1,-1,2,-1.5,rep(0,p-4))
Sigma <- (1/2)^abs(outer(1:p,1:p,FUN = "-"))
chol_Sigma <- chol(Sigma)

# generate X and y
X <- matrix(rnorm(n*p),n,p) \%*\% chol_Sigma
mu <- drop(X \%*\% b)
e <- rnorm(n,0,sigma)
y <- mu + e

# perform least angle regression inference
larinf_out<- larinf(X,y)
plot(larinf_out)

}
\author{
Karl Gregory
}
